DOC: Refactor tsa docs
Add VECM to the docs
Re-org section headers in statespace
Add some minimal docs for VECM
Refactor windows make.bat using latest sphinx

+++ docs/make.bat
@@ -40,7 +40,7 @@ if "%1" == "help" (
 
 if "%1" == "clean" (
 	for /d %%i in (%BUILDDIR%\*) do rmdir /q /s %%i
-	del /q /s %BUILDDIR%\*
+	if exist %BUILDDIR% del /q /s %BUILDDIR%\*
 	goto end
 )
